+++ src/main/java/com/karrot/demo/NumbleKarrotApplication.java
@@ -2,7 +2,9 @@ package com.karrot.demo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cache.annotation.EnableCaching;
 
+@EnableCaching
 @SpringBootApplication
 public class NumbleKarrotApplication {
