+++ src/main/java/com/karrot/demo/redis/CacheValueType.java
@@ -1,0 +1,5 @@
+package com.karrot.demo.redis;
+
+public interface CacheValueType {
+    public String ITEM = "item";
+}
